Refactoring: Risk Level Calculation Error Cases
