+++ src/xcode/ENA/ENA/Source/Services/Risk/Calculation/__tests__/RiskCalculationTests.swift
@@ -229,78 +229,4 @@ class RiskCalculationTests: XCTestCase {
 		XCTAssertEqual(risk?.level, .increased)
 	}
 
-	// MARK: - Risk Level Calculation Error Cases
-
-	func testCalculateRisk_OutsideRangeError_Middle() {
-		let config = RiskProvidingConfiguration(
-			exposureDetectionValidityDuration: .init(day: 1),
-			exposureDetectionInterval: .init(day: 1),
-			detectionMode: .automatic
-		)
-
-		// Test the case where preconditions pass and there is increased risk
-		// Values below are hand-picked to result in a raw risk score of 5.12 - within a gap in the range
-		let summary = makeExposureSummaryContainer(maxRiskScoreFullRange: 128, ad_low: 30, ad_mid: 30, ad_high: 30)
-		let risk = RiskCalculation.risk(
-			summary: summary,
-			configuration: appConfig,
-			// arbitrary, but within limit
-			dateLastExposureDetection: Date().addingTimeInterval(-3600),
-			activeTracing: .init(interval: 48 * 3600),
-			preconditions: preconditions(.valid),
-			previousRiskLevel: nil,
-			providerConfiguration: config
-		)
-
-		XCTAssertNil(risk)
-	}
-
-	func testCalculateRisk_OutsideRangeError_OffHigh() {
-		let config = RiskProvidingConfiguration(
-			exposureDetectionValidityDuration: .init(day: 1),
-			exposureDetectionInterval: .init(day: 1),
-			detectionMode: .automatic
-		)
-
-		// Test the case where preconditions pass and there is increased risk
-		// Values below are hand-picked to result in a raw risk score of 13.6 - outside of the top range of the config
-		let summary = makeExposureSummaryContainer(maxRiskScoreFullRange: 255, ad_low: 40, ad_mid: 40, ad_high: 40)
-		let risk = RiskCalculation.risk(
-			summary: summary,
-			configuration: appConfig,
-			// arbitrary, but within limit
-			dateLastExposureDetection: Date().addingTimeInterval(-3600),
-			activeTracing: .init(interval: 48 * 3600),
-			preconditions: preconditions(.valid),
-			previousRiskLevel: nil,
-			providerConfiguration: config
-		)
-
-		XCTAssertNil(risk)
-	}
-
-	func testCalculateRisk_OutsideRangeError_TooLow() {
-		let config = RiskProvidingConfiguration(
-			exposureDetectionValidityDuration: .init(day: 1),
-			exposureDetectionInterval: .init(day: 1),
-			detectionMode: .automatic
-		)
-
-		// Test the case where preconditions pass and there is increased risk
-		// Values below are hand-picked to result in a raw risk score of 0.85 - outside of the bottom bound of the config
-		let summary = makeExposureSummaryContainer(maxRiskScoreFullRange: 64, ad_low: 10, ad_mid: 10, ad_high: 10)
-		let risk = RiskCalculation.risk(
-			summary: summary,
-			configuration: appConfig,
-			// arbitrary, but within limit
-			dateLastExposureDetection: Date().addingTimeInterval(-3600),
-			activeTracing: .init(interval: 48 * 3600),
-			preconditions: preconditions(.valid),
-			previousRiskLevel: nil,
-			providerConfiguration: config
-		)
-
-		XCTAssertNil(risk)
-	}
-
 }
